Fixes and Additions
Added a small animation (nothing crazy) from Login screen too Home screen. Fixed ForgtPassword so that it goes to the right place. Made it so you can do friend request with a new button to view sent and recieved. Fixed multiple issues of keyboards shunting text inputs around. Set up in-app notification for recieving a friend request and tweaked the alert for gaining/losing coins and xp to a notif that persist for some time.