Read Description
Moved the config folder outside of app to remove a warning. Added some animations for smoother app experience. Fixed up friends screen for the modal and pfps. Fix app navigation in index. Added a loading symbol to ai generation and made it so it can only once at a time. Fixed one thing in shop screen. Fixed colors in settings.